pyproject.toml
[project]
name = "docs2vecs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dorin POMIAN", email = "dorin.pomian@amadeus.com" }
]
requires-python = ">=3.11"
dependencies = [
    "azure-ai-formrecognizer>=3.3.3",
    "azure-search-documents>=11.5.2",
    "azure-storage-blob>=12.24.1",
    "cerberus>=1.3.7",
    "chromadb>=0.6.3",
    "jira>=3.8.0",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-text-splitters>=0.3.6",
    "llama-index>=0.12.20",
    "llama-index-embeddings-azure-openai>=0.3.0",
    "llama-index-embeddings-fastembed>=0.3.0",
    "llama-index-llms-openai>=0.3.20",
    "llama-index-retrievers-bm25>=0.5.2",
    "llama-index-vector-stores-chroma>=0.4.1",
    "markdown>=3.7",
    "openpyxl>=3.1.5",
    "pymongo>=4.11.1",
    "pystemmer>=2.2.0.3",
    "pytest>=8.3.4",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "python-pptx>=1.0.2",
    "unstructured>=0.14.8",
]

[project.scripts]
docs2vecs = "docs2vecs:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
